Move get_dem to dem; punt refactor for now
